Move test to abstract class
